feat: restructure analytics module (#3969)
